PO-2255 - code review changes

+++ src/integrationTest/java/uk/gov/hmcts/opal/service/GenericReportServiceTest.java
@@ -173,8 +173,8 @@ class GenericReportServiceTest extends AbstractIntegrationTest {
         }
 
         @Override
-        public short getNumberOfRecords() {
-            return (short) numberOfRecords;
+        public int getNumberOfRecords() {
+            return numberOfRecords;
         }
 
         @Override
